back/requirements.txt
flask
rsa
pycryptodome
gunicorn
